Fixing up some features

+++ WebProjekat2022/webProjekat2022/src/beans/UserComparatorByPoints.java
@@ -6,6 +6,6 @@ public class UserComparatorByPoints implements Comparator<Customer> {
 @Override
 
 public int compare(Customer o1, Customer o2) {
-        return Long.compare(o1.getPoints(),o2.getPoints());
+        return Double.compare(o1.getPoints(),o2.getPoints());
         }
 }
